streamlit/app/requirements.txt
altair
pandas
streamlit
psycopg2-binary
gremlinpython